Move tests to their respective folders
