+++ Editor/Tests/Nodes/CallTest.cs
@@ -3,21 +3,21 @@ using NUnit.Framework;
 using System;
 using System.Collections.Generic;
 
-namespace Exodrifter.Rumor.Test
+namespace Exodrifter.Rumor.Test.Nodes
 {
 	/// <summary>
-	/// Ensure Jump nodes operate as expected.
+	/// Ensure Call nodes operate as expected.
 	/// </summary>
-	public class JumpTest
+	public class CallTest
 	{
 		/// <summary>
-		/// Ensure jumps to undefined labels throw an exception.
+		/// Ensure calls to undefined labels throw an exception.
 		/// </summary>
 		[Test]
-		public void JumpUndefined()
+		public void CallUndefined()
 		{
-			var rumor = new Engine.Rumor(new List<Node>() {
-				new Jump("start"),
+			var rumor = new Rumor.Engine.Rumor(new List<Node>() {
+				new Call("start"),
 			});
 
 			var yield = rumor.Run();
@@ -25,18 +25,19 @@ namespace Exodrifter.Rumor.Test
 		}
 
 		/// <summary>
-		/// Ensure jumps to defined labels operate as expected.
+		/// Ensure calls to defined labels operate as expected.
 		/// </summary>
 		[Test]
-		public void JumpDefined()
+		public void CallDefined()
 		{
-			var rumor = new Engine.Rumor(new List<Node>() {
-				new Jump("a"),
+			var rumor = new Rumor.Engine.Rumor(new List<Node>() {
+				new Call("a"),
 				new Label("b", null),
 				new Say("b"),
+				new Return(),
 				new Label("a", null),
 				new Say("a"),
-				new Jump("b"),
+				new Call("b"),
 			});
 
 			var yield = rumor.Run();
@@ -46,16 +47,25 @@ namespace Exodrifter.Rumor.Test
 			rumor.Advance();
 			yield.MoveNext();
 			Assert.AreEqual("b", (rumor.Current as Say).text);
+
+			rumor.Advance();
+			yield.MoveNext();
+			Assert.AreEqual("b", (rumor.Current as Say).text);
+
+			rumor.Advance();
+			yield.MoveNext();
+			Assert.True(rumor.Started);
+			Assert.True(rumor.Finished);
 		}
 
 		/// <summary>
-		/// Ensure jumps go to the first defined label when the same label is
+		/// Ensure calls go to the first defined label when the same label is
 		/// defined multiple times in the same scope.
 		/// </summary>
 		[Test]
-		public void JumpMultipleDefinedSameScope()
+		public void CallMultipleDefinedSameScope()
 		{
-			var rumor = new Engine.Rumor(new List<Node>() {
+			var rumor = new Rumor.Engine.Rumor(new List<Node>() {
 				new Jump ("start"),
 				new Label("a", new List<Node>() {
 					new Say("aa"),
@@ -67,7 +77,7 @@ namespace Exodrifter.Rumor.Test
 					new Say("ac"),
 				}),
 				new Label("start", null),
-				new Jump("a"),
+				new Call("a"),
 			});
 
 			var yield = rumor.Run();
@@ -76,13 +86,13 @@ namespace Exodrifter.Rumor.Test
 		}
 
 		/// <summary>
-		/// Ensure jumps go to the closest defined label when the same label
+		/// Ensure calls go to the closest defined label when the same label
 		/// is defined multiple times in different scopes.
 		/// </summary>
 		[Test]
-		public void JumpMultipleDefinedDifferentScope()
+		public void CallMultipleDefinedDifferentScope()
 		{
-			var rumor = new Engine.Rumor(new List<Node>() {
+			var rumor = new Rumor.Engine.Rumor(new List<Node>() {
 				new Label("a", new List<Node>() {
 					new Say("a"),
 
@@ -91,7 +101,7 @@ namespace Exodrifter.Rumor.Test
 					}),
 					new Label("b", new List<Node>() {
 						new Say("ab"),
-						new Jump("a"),
+						new Call("a"),
 					}),
 				}),
 			});
